+++ hybpiper/utils.py
@@ -12,6 +12,7 @@ import scipy
 import textwrap
 from Bio import SeqIO, SeqRecord
 from Bio.Seq import Seq
+import subprocess
 
 
 def log_or_print(string, logger=None, logger_level='info'):
@@ -268,7 +269,6 @@ def translate_target_file(targetfile, logger=None):
     with open(targetfile, 'r') as target_file_handle:
         seqs = list(SeqIO.parse(target_file_handle, 'fasta'))
         for seq in seqs:
-            gene_name = seq.name.split('-')[-1]
             sequence, needed_padding = pad_seq(seq)
             translated_seq = sequence.seq.translate()
             record = SeqRecord.SeqRecord(translated_seq, id=seq.id, description='')
@@ -284,7 +284,61 @@ def translate_target_file(targetfile, logger=None):
 
     with open(f'{translated_target_file}', 'w') as translated_handle:
         SeqIO.write(translated_seqs_to_write, translated_handle, 'fasta')
-    targetfile = translated_target_file
 
-    return targetfile
+    return translated_target_file
+
+
+def check_exonerate_version():
+    """
+    Returns the Exonerate version e.g. 2.4.0
+
+    :return str version: the Exonerate version number
+    """
+
+    exonerate_help = subprocess.run(['exonerate', '-h'], capture_output=True)
+    version = re.search(r'\bexonerate version [0-9][.][0-9].[0-9]\b', str(exonerate_help)).group(0)
+    version_number = version.split()[-1]
+    return version_number
+
+
+def check_dependencies(logger=None):
+    """
+    Checks for the presence of executables and Python packages. Returns a boolean.
+
+    :param logging.Logger logger: a logger object
+    return: bool everything_is_awesome: True if all dependencies are found and are executable, else False
+    """
+
+    executables = ['blastx',
+                   'exonerate',
+                   'parallel',
+                   'makeblastdb',
+                   'spades.py',
+                   'bwa',
+                   'samtools',
+                   'bbmap.sh',
+                   'bbmerge.sh',
+                   'diamond']
+
+    log_or_print(f'{"[INFO]:":10} Checking for external dependencies:\n', logger=logger)
+
+    everything_is_awesome = True
+    for exe in executables:
+        exe_loc = py_which(exe)
+        if exe_loc:
+            log_or_print(f'{exe:20} found at {exe_loc}', logger=logger)
+            if exe == 'exonerate':
+                version = check_exonerate_version()
+                if version not in ['2.4.0']:
+                    everything_is_awesome = False
+
+                    log_or_print(f'\n{exe} version 2.4.0 is required, but your version is {version}. Please update '
+                                 f'your Exonerate version!\n', logger=logger)
+        else:
+            log_or_print(f'{exe:20} not found in your $PATH!', logger=logger)
+            everything_is_awesome = False
+
+    log_or_print('', logger=logger)
+
+    return everything_is_awesome
